refactor: rename package, classes
logger -> aop
LinkoceanAssert -> Assertions
add Assertions.assertThatDataIntegrityViolationException
aop logic
