+++ src/test/java/com/meoguri/linkocean/common/Assertions.java
@@ -3,11 +3,17 @@ package com.meoguri.linkocean.common;
 import static org.assertj.core.api.Assertions.*;
 
 import org.assertj.core.api.ThrowableTypeAssert;
+import org.springframework.dao.DataIntegrityViolationException;
 
 import com.meoguri.linkocean.exception.LinkoceanRuntimeException;
 
-public final class LinkoceanAssert {
+public final class Assertions {
 	public static ThrowableTypeAssert<LinkoceanRuntimeException> assertThatLinkoceanRuntimeException() {
 		return assertThatExceptionOfType(LinkoceanRuntimeException.class);
 	}
+
+	public static ThrowableTypeAssert<DataIntegrityViolationException> assertThatDataIntegrityViolationException() {
+		return assertThatExceptionOfType(DataIntegrityViolationException.class);
+	}
+
 }
