+++ src/test/java/com/meoguri/linkocean/domain/bookmark/entity/BookmarkTest.java
@@ -1,6 +1,6 @@
 package com.meoguri.linkocean.domain.bookmark.entity;
 
-import static com.meoguri.linkocean.common.LinkoceanAssert.*;
+import static com.meoguri.linkocean.common.Assertions.*;
 import static com.meoguri.linkocean.domain.profile.entity.Profile.*;
 import static com.meoguri.linkocean.domain.util.Fixture.*;
 import static java.util.Collections.*;
