+++ src/test/java/com/meoguri/linkocean/domain/bookmark/service/FavoriteServiceImplTest.java
@@ -1,15 +1,14 @@
 package com.meoguri.linkocean.domain.bookmark.service;
 
+import static com.meoguri.linkocean.common.Assertions.*;
 import static com.meoguri.linkocean.domain.util.Fixture.*;
 import static java.util.Collections.*;
-import static org.assertj.core.api.Assertions.assertThatExceptionOfType;
-import static org.assertj.core.api.AssertionsForClassTypes.assertThat;
+import static org.assertj.core.api.AssertionsForClassTypes.*;
 
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.context.SpringBootTest;
-import org.springframework.dao.DataIntegrityViolationException;
 import org.springframework.transaction.annotation.Transactional;
 
 import com.meoguri.linkocean.domain.bookmark.entity.vo.OpenType;
@@ -89,7 +88,7 @@ class FavoriteServiceImplTest {
 		favoriteService.favorite(userId, bookmarkId);
 
 		//when then
-		assertThatExceptionOfType(DataIntegrityViolationException.class)
+		assertThatDataIntegrityViolationException()
 			.isThrownBy(() -> favoriteService.favorite(userId, bookmarkId));
 	}
