+++ src/test/java/com/meoguri/linkocean/domain/profile/entity/FollowTest.java
@@ -1,6 +1,6 @@
 package com.meoguri.linkocean.domain.profile.entity;
 
-import static com.meoguri.linkocean.common.LinkoceanAssert.*;
+import static com.meoguri.linkocean.common.Assertions.*;
 import static com.meoguri.linkocean.domain.util.Fixture.*;
 
 import org.junit.jupiter.api.Test;
